requirements.txt
tensorflow>=2.12.0
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
Pillow>=9.4.0
pyyaml>=6.0
tqdm>=4.65.0
